refactor(net): extract `ipv4` and `ipv6` modules into separate src files
